[Refactor] #10 - 가격 Integer로 수정 및 도메인 내 검증 로직 추가

+++ baro-buyer/src/main/java/com/barofarm/buyer/product/application/dto/ProductCreateCommand.java
@@ -1,7 +1,6 @@
 package com.barofarm.buyer.product.application.dto;
 
 import com.barofarm.buyer.product.domain.ProductCategory;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductCreateCommand(
@@ -10,5 +9,5 @@ public record ProductCreateCommand(
     String productName,
     String description,
     ProductCategory productCategory,
-    BigDecimal price,
+    Integer price,
     Integer stockQuantity) {}
